refactor : API 명세에 따른 타입 변경 #PAR-380

+++ core/model/src/main/java/online/partyrun/partyrunapplication/core/model/my_page/ComprehensiveRunRecord.kt
@@ -2,6 +2,6 @@ package online.partyrun.partyrunapplication.core.model.my_page
 
 data class ComprehensiveRunRecord(
     val averagePace: Double,
-    val totalDistance: Int,
+    val totalDistance: Double,
     val totalRunningTime: TotalRunningTime
 )
